+++ CHANGELOG.md
@@ -53,3 +53,4 @@
 # 0.0.3
 * Make `Unit` type family closed
 * Define `-=-` and `=-=` operators in `Disbandable` typeclass
+* Define `|->` operator in `Adj.Algebra.Category` module