+++ CHANGELOG.md
@@ -54,3 +54,4 @@
 * Make `Unit` type family closed
 * Define `-=-` and `=-=` operators in `Disbandable` typeclass
 * Define `|->` operator in `Adj.Algebra.Category` module
+* Define `Betwixt` type family in `Adj.Algebara.Category` module