+++ CHANGELOG.md
@@ -65,3 +65,4 @@
 * Move `Casting` typeclass to `Adj.Auxiliary` method
 * Define `Tensor` morpism
 * Define `Component` typeclass
+* Define `Transformation` typeclass